Split Mirror from MirrorRepository.
