+++ src/plugins/discourse/mirror.test.js
@@ -2,9 +2,8 @@
 
 import sortBy from "lodash.sortby";
 import Database from "better-sqlite3";
-import fs from "fs";
-import tmp from "tmp";
 import {Mirror} from "./mirror";
+import {SqliteMirrorRepository} from "./mirrorRepository";
 import {
   type Discourse,
   type TopicId,
@@ -170,60 +169,41 @@ describe("plugins/discourse/mirror", () => {
     const fetcher = new MockFetcher();
     const db = new Database(":memory:");
     const url = "http://example.com";
-    const mirror = new Mirror(db, fetcher, url);
+    const repo = new SqliteMirrorRepository(db, url);
+    const mirror = new Mirror(repo, fetcher, url);
     const reporter = new TestTaskReporter();
-    return {fetcher, mirror, reporter, url};
+    return {fetcher, mirror, reporter, url, repo};
   };
 
-  it("rejects a different server url without changing the database", () => {
-    // We use an on-disk database file here so that we can dump the
-    // contents to ensure that the database is physically unchanged.
-    const filename = tmp.fileSync().name;
-    const db = new Database(filename);
-    const fetcher = new MockFetcher();
-    const url1 = "https://foo.bar";
-    const url2 = "https://foo.zod";
-    expect(() => new Mirror(db, fetcher, url1)).not.toThrow();
-    const data = fs.readFileSync(filename).toJSON();
-
-    expect(() => new Mirror(db, fetcher, url2)).toThrow(
-      "incompatible server or version"
-    );
-    expect(fs.readFileSync(filename).toJSON()).toEqual(data);
-
-    expect(() => new Mirror(db, fetcher, url1)).not.toThrow();
-    expect(fs.readFileSync(filename).toJSON()).toEqual(data);
-  });
-
   it("mirrors topics from the fetcher", async () => {
-    const {mirror, fetcher, reporter} = example();
+    const {mirror, fetcher, reporter, repo} = example();
     fetcher.addPost(2, null);
     fetcher.addPost(3, null);
     const topic2 = fetcher._topic(2);
     const topic3 = fetcher._topic(3);
     await mirror.update(reporter);
-    expect(mirror.topics()).toEqual([topic2, topic3]);
+    expect(repo.topics()).toEqual([topic2, topic3]);
   });
 
   it("mirrors posts from the fetcher", async () => {
-    const {mirror, fetcher, reporter} = example();
+    const {mirror, fetcher, reporter, repo} = example();
     const p1 = fetcher.addPost(2, null);
     const p2 = fetcher.addPost(3, null);
     const p3 = fetcher.addPost(3, 1);
     await mirror.update(reporter);
     const posts = [fetcher._post(p1), fetcher._post(p2), fetcher._post(p3)];
-    expect(mirror.posts()).toEqual(posts);
+    expect(repo.posts()).toEqual(posts);
   });
 
   it("provides usernames for all active users", async () => {
-    const {mirror, fetcher, reporter} = example();
+    const {mirror, fetcher, reporter, repo} = example();
     fetcher.addPost(2, null, "alpha");
     fetcher.addPost(3, null, "beta");
     fetcher.addPost(3, 1, "alpha");
     await mirror.update(reporter);
     // credbot appears because it is the nominal author of all topics
     expect(
-      mirror
+      repo
         .users()
         .slice()
         .sort()
@@ -236,7 +216,7 @@ describe("plugins/discourse/mirror", () => {
   }
 
   it("provides all the likes by users that have posted", async () => {
-    const {mirror, fetcher, reporter} = example();
+    const {mirror, fetcher, reporter, repo} = example();
     fetcher.addPost(1, null, "alpha");
     fetcher.addPost(2, null, "alpha");
     fetcher.addPost(3, null, "beta");
@@ -245,26 +225,26 @@ describe("plugins/discourse/mirror", () => {
     const l3 = fetcher.addLike("beta", 3, 7);
     const l4 = fetcher.addLike("alpha", 1, 8);
     await mirror.update(reporter);
-    expectLikesSorted(mirror.likes(), [l1, l2, l3, l4]);
+    expectLikesSorted(repo.likes(), [l1, l2, l3, l4]);
     const l5 = fetcher.addLike("alpha", 2, 9);
     fetcher.addPost(4, null, "credbot");
     const l6 = fetcher.addLike("credbot", 2, 10);
     const l7 = fetcher.addLike("beta", 4, 11);
     await mirror.update(reporter);
-    expectLikesSorted(mirror.likes(), [l1, l2, l3, l4, l5, l6, l7]);
+    expectLikesSorted(repo.likes(), [l1, l2, l3, l4, l5, l6, l7]);
   });
 
   it("doesn't find likes of users that never posted", async () => {
-    const {mirror, fetcher, reporter} = example();
+    const {mirror, fetcher, reporter, repo} = example();
     fetcher.addPost(1, null);
     fetcher.addLike("nope", 1, 1);
     await mirror.update(reporter);
-    expect(mirror.likes()).toEqual([]);
+    expect(repo.likes()).toEqual([]);
   });
 
   describe("update semantics", () => {
     it("only fetches new topics on `update`", async () => {
-      const {mirror, fetcher, reporter} = example();
+      const {mirror, fetcher, reporter, repo} = example();
       fetcher.addPost(1, null);
       fetcher.addPost(2, null);
       await mirror.update(reporter);
@@ -273,11 +253,11 @@ describe("plugins/discourse/mirror", () => {
       await mirror.update(reporter);
       expect(fetchTopicWithPosts).toHaveBeenCalledTimes(1);
       expect(fetchTopicWithPosts).toHaveBeenCalledWith(3);
-      expect(mirror.topics().map((x) => x.id)).toEqual([1, 2, 3]);
+      expect(repo.topics().map((x) => x.id)).toEqual([1, 2, 3]);
     });
 
     it("gets new posts on old topics on update", async () => {
-      const {mirror, fetcher, reporter} = example();
+      const {mirror, fetcher, reporter, repo} = example();
       fetcher.addPost(1, null);
       fetcher.addPost(2, null);
       await mirror.update(reporter);
@@ -287,30 +267,30 @@ describe("plugins/discourse/mirror", () => {
       const latestPosts = await fetcher.latestPosts();
       // The post added to the old topic wasn't retrieved by latest post
       expect(latestPosts.map((x) => x.id)).not.toContain(id);
-      const allPostIds = mirror.posts().map((x) => x.id);
+      const allPostIds = repo.posts().map((x) => x.id);
       // The post was still included, meaning the mirror scanned for new posts by id
       expect(allPostIds).toContain(id);
     });
 
     it("skips null/missing topics", async () => {
-      const {mirror, fetcher, reporter} = example();
+      const {mirror, fetcher, reporter, repo} = example();
       fetcher.addPost(1, null);
       fetcher.addPost(3, null);
       await mirror.update(reporter);
-      expect(mirror.topics().map((x) => x.id)).toEqual([1, 3]);
+      expect(repo.topics().map((x) => x.id)).toEqual([1, 3]);
     });
 
     it("skips null/missing posts", async () => {
-      const {mirror, fetcher, reporter} = example();
+      const {mirror, fetcher, reporter, repo} = example();
       const p1 = fetcher.addPost(1, null);
       fetcher._latestPostId += 2;
       const p2 = fetcher.addPost(3, null);
       await mirror.update(reporter);
-      expect(mirror.posts().map((x) => x.id)).toEqual([p1, p2]);
+      expect(repo.posts().map((x) => x.id)).toEqual([p1, p2]);
     });
 
     it("queries explicitly for posts that are not present in topicWithPosts.posts", async () => {
-      const {mirror, fetcher, reporter} = example();
+      const {mirror, fetcher, reporter, repo} = example();
       const p1 = fetcher.addPost(1, null);
       const p2 = fetcher.addPost(1, 1);
       const p3 = fetcher.addPost(1, 1);
@@ -328,7 +308,7 @@ describe("plugins/discourse/mirror", () => {
       expect(fetchPost).toHaveBeenCalledTimes(1);
       expect(fetchPost).toHaveBeenCalledWith(p2);
 
-      expect(mirror.posts().map(getId)).toEqual([p1, p2, p3]);
+      expect(repo.posts().map(getId)).toEqual([p1, p2, p3]);
     });
 
     it("does not explicitly query for posts that were in topicWithPosts.posts", async () => {
@@ -340,14 +320,14 @@ describe("plugins/discourse/mirror", () => {
     });
 
     it("does not explicitly query for posts that were provided in latest posts", async () => {
-      const {mirror, fetcher, reporter} = example();
+      const {mirror, fetcher, reporter, repo} = example();
       fetcher.addPost(1, null);
       await mirror.update(reporter);
       const id = fetcher.addPost(1, 1);
       const fetchPost = jest.spyOn(fetcher, "post");
       await mirror.update(reporter);
       expect(fetchPost).not.toHaveBeenCalled();
-      expect(mirror.posts().map((x) => x.id)).toContain(id);
+      expect(repo.posts().map((x) => x.id)).toContain(id);
     });
 
     it("does not query for topics at all if there were no new topics", async () => {
@@ -400,7 +380,7 @@ describe("plugins/discourse/mirror", () => {
     });
 
     it("warns if one of the latest posts has no topic", async () => {
-      const {mirror, fetcher, reporter} = example();
+      const {mirror, fetcher, reporter, repo} = example();
       const pid1 = fetcher.addPost(1, null, "credbot");
       const pid2 = fetcher.addPost(2, null, "credbot");
       // Verify that the problem post is one of the latest posts
@@ -410,9 +390,9 @@ describe("plugins/discourse/mirror", () => {
       fetcher._latestTopicId--;
       await mirror.update(reporter);
       const topics = [fetcher._topic(1)];
-      expect(mirror.topics()).toEqual(topics);
+      expect(repo.topics()).toEqual(topics);
       const posts = [fetcher._post(pid1)];
-      expect(mirror.posts()).toEqual(posts);
+      expect(repo.posts()).toEqual(posts);
       expect(console.warn).toHaveBeenCalledWith(
         "Warning: Encountered error 'FOREIGN KEY constraint failed' " +
           "while adding post http://example.com/t/2/1."
@@ -422,7 +402,7 @@ describe("plugins/discourse/mirror", () => {
     });
 
     it("warns if it finds a (non-latest) post with no topic", async () => {
-      const {mirror, fetcher, reporter} = example();
+      const {mirror, fetcher, reporter, repo} = example();
       const pid1 = fetcher.addPost(1, null, "credbot");
       const pid2 = fetcher.addPost(2, null, "credbot");
       const pid3 = fetcher.addPost(1, null, "credbot");
@@ -433,9 +413,9 @@ describe("plugins/discourse/mirror", () => {
       fetcher._latestTopicId--;
       await mirror.update(reporter);
       const topics = [fetcher._topic(1)];
-      expect(mirror.topics()).toEqual(topics);
+      expect(repo.topics()).toEqual(topics);
       const posts = [pid1, pid3].map((x) => fetcher._post(x));
-      expect(mirror.posts()).toEqual(posts);
+      expect(repo.posts()).toEqual(posts);
       expect(console.warn).toHaveBeenCalledWith(
         "Warning: Encountered error 'FOREIGN KEY constraint failed' " +
           "while adding post http://example.com/t/2/1."
@@ -445,14 +425,14 @@ describe("plugins/discourse/mirror", () => {
     });
 
     it("warns if it gets a like that doesn't correspond to any post", async () => {
-      const {mirror, fetcher, reporter} = example();
+      const {mirror, fetcher, reporter, repo} = example();
       const pid = fetcher.addPost(1, null, "credbot");
       const badLike = {username: "credbot", postId: 37, timestampMs: 0};
       fetcher._likes.push(badLike);
       await mirror.update(reporter);
-      expect(mirror.topics()).toEqual([fetcher._topic(1)]);
-      expect(mirror.posts()).toEqual([fetcher._post(pid)]);
-      expect(mirror.likes()).toEqual([]);
+      expect(repo.topics()).toEqual([fetcher._topic(1)]);
+      expect(repo.posts()).toEqual([fetcher._post(pid)]);
+      expect(repo.likes()).toEqual([]);
       expect(console.warn).toHaveBeenCalledWith(
         "Warning: Encountered error 'FOREIGN KEY constraint failed' " +
           "on a like by credbot on post id 37."
@@ -481,42 +461,42 @@ describe("plugins/discourse/mirror", () => {
 
   describe("findPostInTopic", () => {
     it("works for the first post in a topic", async () => {
-      const {mirror, fetcher, reporter} = example();
+      const {mirror, fetcher, reporter, repo} = example();
       const id = fetcher.addPost(5, null);
       const post = NullUtil.get(fetcher._post(id));
       expect(post.topicId).toEqual(5);
       expect(post.indexWithinTopic).toEqual(1);
       await mirror.update(reporter);
-      expect(mirror.findPostInTopic(5, 1)).toEqual(id);
+      expect(repo.findPostInTopic(5, 1)).toEqual(id);
     });
 
     it("works for the second post in a topic", async () => {
-      const {mirror, fetcher, reporter} = example();
+      const {mirror, fetcher, reporter, repo} = example();
       fetcher.addPost(1, null);
       const id = fetcher.addPost(1, 1);
       const post = NullUtil.get(fetcher._post(id));
       expect(post.indexWithinTopic).toEqual(2);
       await mirror.update(reporter);
-      expect(mirror.findPostInTopic(1, 2)).toEqual(id);
+      expect(repo.findPostInTopic(1, 2)).toEqual(id);
     });
 
     it("returns undefined for a post with too high an index", async () => {
-      const {mirror, fetcher, reporter} = example();
+      const {mirror, fetcher, reporter, repo} = example();
       fetcher.addPost(1, null);
       await mirror.update(reporter);
-      expect(mirror.findPostInTopic(1, 2)).toBe(undefined);
+      expect(repo.findPostInTopic(1, 2)).toBe(undefined);
     });
 
     it("returns undefined for topic that doesnt exist", async () => {
-      const {mirror, fetcher, reporter} = example();
+      const {mirror, fetcher, reporter, repo} = example();
       fetcher.addPost(1, null);
       await mirror.update(reporter);
-      expect(mirror.findPostInTopic(2, 1)).toBe(undefined);
+      expect(repo.findPostInTopic(2, 1)).toBe(undefined);
     });
 
     it("returns undefined for a mirror that never updated", async () => {
-      const {mirror} = example();
-      expect(mirror.findPostInTopic(1, 1)).toBe(undefined);
+      const {repo} = example();
+      expect(repo.findPostInTopic(1, 1)).toBe(undefined);
     });
   });
 });
