+++ src/plugins/discourse/mirrorRepository.test.js
@@ -1,0 +1,27 @@
+// @flow
+
+import Database from "better-sqlite3";
+import fs from "fs";
+import tmp from "tmp";
+import {SqliteMirrorRepository} from "./mirrorRepository";
+
+describe("plugins/discourse/mirrorRepository", () => {
+  it("rejects a different server url without changing the database", () => {
+    // We use an on-disk database file here so that we can dump the
+    // contents to ensure that the database is physically unchanged.
+    const filename = tmp.fileSync().name;
+    const db = new Database(filename);
+    const url1 = "https://foo.bar";
+    const url2 = "https://foo.zod";
+    expect(() => new SqliteMirrorRepository(db, url1)).not.toThrow();
+    const data = fs.readFileSync(filename).toJSON();
+
+    expect(() => new SqliteMirrorRepository(db, url2)).toThrow(
+      "incompatible server or version"
+    );
+    expect(fs.readFileSync(filename).toJSON()).toEqual(data);
+
+    expect(() => new SqliteMirrorRepository(db, url1)).not.toThrow();
+    expect(fs.readFileSync(filename).toJSON()).toEqual(data);
+  });
+});
